requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
sentence-transformers
transformers
langchain
huggingface_hub
chromadb
rankbm25
PyPDF2